Extract useSpatialSearch hook
