+++ map/src/util/hooks/search/useSpatialSearch.js
@@ -1,0 +1,33 @@
+import { useCallback, useContext } from 'react';
+import { useLocation, useNavigate } from 'react-router-dom';
+import AppContext, {
+    SEARCH_ENGINE_CLASSIC,
+    SEARCH_ENGINE_SPATIAL,
+    SPATIAL_SEARCH_STORAGE_KEY,
+} from '../../../context/AppContext';
+import { liveHash } from '../../../manager/GlobalManager';
+
+const ENGINE_KEY = 'engine';
+
+export function engineFromSpatial(on) {
+    return on ? SEARCH_ENGINE_SPATIAL : SEARCH_ENGINE_CLASSIC;
+}
+
+export default function useSpatialSearch() {
+    const ctx = useContext(AppContext);
+    const location = useLocation();
+    const navigate = useNavigate();
+
+    const setSpatial = useCallback(
+        (on) => {
+            ctx.setSpatialSearch(on);
+            localStorage.setItem(SPATIAL_SEARCH_STORAGE_KEY, on ? 'yes' : 'no');
+            const search = new URLSearchParams(location.search);
+            search.set(ENGINE_KEY, engineFromSpatial(on));
+            navigate({ pathname: location.pathname, search: `?${search}`, hash: liveHash() });
+        },
+        [ctx, location, navigate]
+    );
+
+    return { spatial: ctx.spatialSearch, setSpatial };
+}
